requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
jinja2>=3.1.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
sqlalchemy>=2.0.23
alembic>=1.12.1
pydantic>=2.6.0
qrcode[pil]>=7.4.2
pillow>=10.2.0
aiofiles>=23.2.1
python-dotenv>=1.0.0
reportlab>=4.0.7
weasyprint>=60.2
